module: crossover: Unify sink count type and change frames parameter type
Unify sink count types to uint32_t as ipc configuration use it. Use size_t
for frame counts in processing functions.

+++ src/audio/crossover/crossover.h
@@ -57,8 +57,8 @@ struct comp_data;
 typedef int (*crossover_process_func)(struct comp_data *cd,
 				      struct sof_source *source,
 				      struct sof_sink **sinks,
-				      int32_t num_sinks,
-				      uint32_t frames);
+				      uint32_t num_sinks,
+				      size_t frames);
 
 /* Crossover component private data */
 struct comp_data {
